Review

+++ src/weathergen/datasets/stream_data.py
@@ -106,7 +106,7 @@ class StreamData:
         """
 
         self.source_raw += [torch.tensor([])]
-        self.source_tokens_lens += [torch.zeros([self.nhc_source], dtype=torch.int64)]
+        self.source_tokens_lens += [torch.zeros([self.nhc_source], dtype=torch.int32)]
         self.source_tokens_cells += [torch.tensor([])]
         self.source_centroids += [torch.tensor([])]
 
@@ -124,12 +124,10 @@ class StreamData:
         None
         """
 
-        # self.target_tokens[fstep] += [
-        #     [torch.zeros((0,0), dtype=torch.int64) for _ in range(self.nhc_target)]]
-        self.target_tokens[fstep] += [[torch.tensor([], dtype=torch.int64)]]
-        self.target_tokens_lens[fstep] += [torch.zeros([self.nhc_target], dtype=torch.int64)]
+        self.target_tokens[fstep] += [[torch.tensor([], dtype=torch.int32)]]
+        self.target_tokens_lens[fstep] += [torch.zeros([self.nhc_target], dtype=torch.int32)]
         self.target_coords[fstep] += [[torch.zeros((0, 106)) for _ in range(self.nhc_target)]]
-        self.target_coords_lens[fstep] += [torch.zeros([self.nhc_target], dtype=torch.int64)]
+        self.target_coords_lens[fstep] += [torch.zeros([self.nhc_target], dtype=torch.int32)]
         self.target_coords_raw[fstep] += [[torch.tensor([]) for _ in range(self.nhc_target)]]
         self.target_times_raw[fstep] += [
             [np.array([], dtype="datetime64[ns]") for _ in range(self.nhc_target)]
